+++ app/src/main/java/com/nemodream/bangkkujaengi/customer/ui/fragment/socialChildFragment/SocialFollowingFragment.kt
@@ -1,4 +1,4 @@
-package com.nemodream.bangkkujaengi.customer.ui.fragment
+package com.nemodream.bangkkujaengi.customer.ui.fragment.socialChildFragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [SocialMyFragment.newInstance] factory method to
+ * Use the [SocialFollowingFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class SocialMyFragment : Fragment() {
+class SocialFollowingFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class SocialMyFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_social_my, container, false)
+        return inflater.inflate(R.layout.fragment_social_following, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class SocialMyFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment SocialMyFragment.
+         * @return A new instance of fragment SocialFollowingFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            SocialMyFragment().apply {
+            SocialFollowingFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
