+++ app/src/main/java/com/nemodream/bangkkujaengi/customer/ui/fragment/socialChildFragment/SocialRankFragment.kt
@@ -1,4 +1,4 @@
-package com.nemodream.bangkkujaengi.customer.ui.fragment
+package com.nemodream.bangkkujaengi.customer.ui.fragment.socialChildFragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [SocialFollowingFragment.newInstance] factory method to
+ * Use the [SocialRankFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class SocialFollowingFragment : Fragment() {
+class SocialRankFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class SocialFollowingFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_social_following, container, false)
+        return inflater.inflate(R.layout.fragment_social_rank, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class SocialFollowingFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment SocialFollowingFragment.
+         * @return A new instance of fragment SocialRankFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            SocialFollowingFragment().apply {
+            SocialRankFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
