+++ app/src/main/res/layout/fragment_social_rank.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".customer.ui.fragment.SocialRankFragment">
+    tools:context=".customer.ui.fragment.socialChildFragment.SocialRankFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
